Diagram and table updates
Updated the 'D-Comprar-material-de-lista-basica-AF-teste-VA' diagram with the payment register tasks;

Updated table T2 including the fields related to the payment registration (orderReceptPayDate,orderReceptPayNote);

Updated the JSon file T2 metadata with the description of the new fields.

Updated the SQL to insertion of the updated data related to the Supply Athorization (needs verification).